requirements.txt
collections
copy
glob
importlib
matplotlib
mpl_toolkits
numpy
numpy-stl
os
pandas
pickle
PIL
pyvista
scipy
sklearn
sys
tifffile
time
tkinter
torch
vtk
